Test de Funkos creados

+++ src/main/java/com/example/springfunko/category/services/CategoryServiceImpl.java
@@ -1,13 +1,17 @@
 package com.example.springfunko.category.services;
 
 import com.example.springfunko.category.dto.CategoryResponseDto;
+import com.example.springfunko.category.exception.CategoryConflict;
 import com.example.springfunko.category.exception.CategoryNotFound;
 import com.example.springfunko.category.mappers.CategoryMapper;
 import com.example.springfunko.category.models.Categoria;
 import com.example.springfunko.category.repositories.CategoryRepository;
+import jakarta.transaction.Transactional;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.cache.annotation.CacheConfig;
+import org.springframework.cache.annotation.CacheEvict;
+import org.springframework.cache.annotation.Cacheable;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -38,12 +42,14 @@ public class CategoryServiceImpl implements CategoryService {
     }
 
     @Override
+    @Cacheable(key = "#id")
     public Categoria findById(Long id) {
         log.info("Buscando categoria por id");
         return categoryRepository.findById(id).orElseThrow(() -> new CategoryNotFound("Categoria no encontrada"));
     }
 
     @Override
+    @Cacheable(key = "#result.id")
     public Categoria save(CategoryResponseDto categoria) {
         log.info("Guardando categoria");
         Categoria categoriaMapped = categoryMapper.toCategory(categoria);
@@ -51,6 +57,7 @@ public class CategoryServiceImpl implements CategoryService {
     }
 
     @Override
+    @Cacheable(key = "#result.id")
     public Categoria update(CategoryResponseDto categoria, Long id) {
         log.info("Actualizando categoria");
         Categoria categoryActual = findById(id);
@@ -58,8 +65,15 @@ public class CategoryServiceImpl implements CategoryService {
     }
 
     @Override
+    @CacheEvict(key = "#id")
+    @Transactional
     public void deleteById(Long id) {
         log.info("Eliminando categoria");
-        categoryRepository.deleteById(id);
+        Categoria categoria = findById(id);
+        if (categoryRepository.existsFunkoById(id)){
+            throw new CategoryConflict("Categoria no puede ser eliminada porque tiene funkos asociados");
+        } else {
+            categoryRepository.deleteById(id);
+        }
     }
 }
